requirements/constraints.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==2.2.16
django-bootstrap3==14.1.0
django-crispy-forms==1.9.2
django-dbmi-client==0.3.16
django-health-check==3.12.1
djangorestframework==3.11.1
fhirclient==3.2.0
furl==2.1.0
idna==2.10
importlib-metadata==1.7.0
isodate==0.6.0
jwcrypto==0.7
orderedmultidict==1.0.1
ppm-utils==0.9.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
raven==6.10.0
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
